backend/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.29
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4

